all done

+++ routes/users.js
@@ -105,7 +105,7 @@ router
       if (user) {
         req.session.user = createSessionObject(user);
       }
-      return res.status(200).redirect("/calendar/month");
+      return res.redirect("/calendar/month");
     } catch (e) {
       return res.status(404).render("user/signup", {
         title: "Sign Up",
@@ -115,7 +115,7 @@ router
     }
   })
   .get(async (req, res) => {
-    if (req.session.user) {
+    if (req?.session?.user) {
       return res.redirect("/calendar/month");
     }
     res.render("user/signup", {
@@ -126,15 +126,12 @@ router
 router
   .route("/login")
   .post(async (req, res) => {
-    if (req.session.user) {
+    if (req?.session?.user) {
       return res.redirect("/calendar/month");
     }
     let email = req.body.email;
     let password = req.body.password;
 
-    if (req.session.user) {
-      return res.redirect("/calendar");
-    }
     let errorMessages = {};
     try {
       utils.validateEmail(email);
@@ -184,39 +181,46 @@ router
     }
   })
   .get(async (req, res) => {
-    if (req.session.user) {
+    if (req?.session?.user) {
       return res.redirect("/calendar/month");
     }
     res.render("user/login", {
       title: "Login",
     });
   });
-router
-  .route("/profile")
 
-  .get(async (req, res) => {
-    try {
-      if (req.session.user) {
-        const id = req.session.user.user_id;
-        const currUser = await usersFunctions.getUser(id);
+function validateUser(req, res, next) {
+  if (!req?.session?.user) {
+    return res.status(403).render("errors/error", {
+      title: "Error",
+      error: new Error("HTTP Error 403 : please Login"),
+    });
+  }
+  next();
+}
+router.route("/profile").get(validateUser, async (req, res) => {
+  try {
+    if (req?.session?.user) {
+      const id = req.session.user.user_id;
+      const currUser = await usersFunctions.getUser(id);
 
-        return res.render("user/profile", {
-          title: "Profile",
-          currUser,
-        });
-      } else {
-        return res.render("user/login", { title: "Login" });
-      }
-    } catch (e) {
-      return res.status(500).json({ error: e.message });
+      return res.render("user/profile", {
+        title: "Profile",
+        currUser,
+      });
+    } else {
+      return res.render("user/login", { title: "Login" });
     }
-  });
+  } catch (e) {
+    return res.status(500).json({ error: e.message });
+  }
+});
 
 router
   .route("/edit")
-  .get(async (req, res) => {
-    if (!req.session.user) {
-      res.render("user/login", { title: "Login" });
+  .get(validateUser, async (req, res) => {
+    if (!req?.session?.user) {
+      res.status(403).render("user/login", { title: "Login" });
     } else {
       const id = req.session.user.user_id;
       try {
@@ -229,7 +233,7 @@ router
     }
   })
   .post(async (req, res) => {
-    if (!req.session.user) {
+    if (!req?.session?.user) {
       res.render("user/login", { title: "Login" });
     } else {
       const id = req.session.user.user_id;
@@ -291,14 +295,10 @@ router
   });
 router
   .route("/password")
-  .get(async (req, res) => {
-    if (req.session.user) {
-      return res.render("user/password", { title: "Password" });
-    }
-
-    res.redirect("/");
+  .get(validateUser, async (req, res) => {
+    return res.render("user/password", { title: "Password" });
   })
-  .post(async (req, res) => {
+  .post(validateUser, async (req, res) => {
     let errorMessages = {};
 
     if (!req?.session?.user || !req.session.user.user_id) {
@@ -376,7 +376,7 @@ router
         req.body.newPassword,
         req.body.reEnterNewPassword
       );
-      return res.status(200).redirect("user/password", {
+      return res.redirect("user/password", {
         success: "password successfully changed",
       });
     } catch (e) {
@@ -387,25 +387,52 @@ router
       });
     }
   });
-router.route("/deleteEvents").get(async (req, res) => {
-  const id = req.session.user.user_id;
-  await usersFunctions.deleteAllEvents(id);
-  res.redirect("/calendar/month");
-});
-router.route("/deleteUser").delete(async (req, res) => {
-  const id = req.session.user.user_id;
-  await usersFunctions.deleteUser(id);
+router
+  .route("/deleteEvents/:confirmDelete")
+  .delete(validateUser, async (req, res) => {
+    const id = req.session.user.user_id;
+    const confirmDelete = utils.validateBooleanInput(
+      req?.params?.confirmDelete
+    );
+    if (confirmDelete) {
+      try {
+        await usersFunctions.deleteAllEvents(id);
+
+        return res.status(200).json({ message: "All Events Deleted" });
+      } catch (error) {
+        return res.status(500).render("errors/error", {
+          title: "Error",
+          error: new Error("Something went wrong. Please try again later"),
+        });
+      }
+    } else {
+      res.status(403).json({ error: "No confirmation was given" });
+    }
+  });
+router
+  .route("/deleteUser/:confirmDelete")
+  .get(validateUser, async (req, res) => {
+    const id = req.session.user.user_id;
+    const confirmDelete = utils.validateBooleanInput(
+      req?.params?.confirmDelete
+    );
+    if (confirmDelete) {
+      try {
+        await usersFunctions.deleteUser(id);
+        res.clearCookie("AuthCookie");
+        req.session.destroy();
+        return res.redirect("/user/login");
+      } catch (error) {
+        return res.status(500).render("errors/error", {
+          title: "Error",
+          error: new Error("Something went wrong. Please try again later"),
+        });
+      }
+    }
+  });
+router.route("/logout").get(validateUser, async (req, res) => {
   res.clearCookie("AuthCookie");
   req.session.destroy();
-  res.status(200).json("Successful");
-  //res.redirect("/user/login");
-});
-router.route("/logout").get(async (req, res) => {
-  if (!req.session.user) {
-    return res.redirect("/user/login");
-  }
-  res.clearCookie("AuthCookie");
-  req.session.destroy();
-  res.redirect("/user/login");
+  return res.redirect("/user/login");
 });
 export default router;
